start hacking on a new tool: worklog
Most of the design work right now lives in README.md and Cargo.toml,
where I'm collecting ideas for how to do this. I'm tentative on using
Diesel, which seems plausible but also seems like it might be
inordinately heavy for what I'm doing here. Also it is not at all
obvious that its struct-oriented approach will work at all with the
potentially wacky SQL I'm going to need to write to handle the
time-tracking portions of the task. Still, might as well get used
to the library; we are in fact using it at work.