WIP: try using a custom command parser instead of clap/structopt
Given the style of command line I'm using, and the relative simplicity
of the help messages I'll be writing, a custom parser seems like
it _might_ be the way to go here.

LALRPOP is still undecided though. I have some experience in it,
and it's kind of frustrating. Might be better to switch to Pest
<https://pest.rs>.