try rewriting the parser in pest
On the one hand, it kind of looks like it's a relatively powerful
parser, and it does what I want without too much pain, so that's good.
On the other hand, that's a whole lot of `.unwrap()` in the example!
Need to do some research to ensure that we're doing it right.