fixup the parser error handling
Turns out that I was thinking about error handling wrong. While it
is possible in principle to depend on the PEG grammar parse errors
to see what's going on, that's not supported. And that's no bad thing!
Because it pushes us to do the smarter thing, which is to return
Result<T, Error> from the relevant sections of the grammar, and get
nice semantic errors just how we'd expect them.