insert events using raw sql
Yeah, sqlx was the simpler solution here. It's turning out to be
relatively straightforward, even the sneaky little transform from
`DateTime<Local>` outside the `db` module to `DateTime<Utc>`
within it.

Cursory checking seems to indicate that there's a silent failure
somewhere in the data pipeline; it looks like the DB is not actually
being written to when I run the CLI. That'll need to be investigate.

That's for later, though! For now, celebrate basic functionality.