log universal row id, not per day. force insert to flush.
Logging a non-transient row id will make things easier in case we ever
want to remove or edit a row. It also means that when we get to reports,
we can see what was inserted out of order.

We also force inserts to flush. We'd previously been just executing
the insert and then expecting the drop handler to finalize it. This
happened... sometimes. What's way more effective is to create a
transaction, and explicitly commit that transaction. This is what
we do.